requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
protobuf==3.20.1
python-dotenv==0.21.0
Werkzeug==2.2.2
gunicorn==20.1.0